+++ tests/testthat/test-http-httr2.R
@@ -1,8 +1,5 @@
-test_that("basic HTTP methods work", {
-  skip_if_not_installed("RCurl")
+test_that("basic HTTP methods work (httr2)", {
   skip_if_not_installed("webfakes")
-
-  withr::local_options(rsconnect.http = "rcurl", lifecycle_verbosity = "quiet")
 
   test_http_GET()
   test_http_POST_JSON()