+++ tests/testthat/test-http-libcurl.R
@@ -1,7 +1,7 @@
+withr::local_options(rsconnect.httr2 = FALSE)
+
 test_that("basic HTTP methods work", {
   skip_if_not_installed("webfakes")
-
-  withr::local_options(rsconnect.http = "libcurl")
 
   test_http_GET()
   test_http_POST_JSON()