+++ NEWS.md
@@ -1,12 +1,57 @@
+# rsconnect 1.8.0
+
+* `rsconnect` now uses [`httr2`](https://httr2.r-lib.org/) as its HTTP client.
+  There should be no user-visible changes as a result, but if something does
+  not work as expected, please file an issue, and you can set
+  `options(rsconnect.httr2 = FALSE)` as a temporary workaround. (#1284)
+
+* Added support for renv profiles and renv lockfiles that are located outside of
+  the project root. (#1122)
+
+* Resolved a bug where `renv.lock` files that had multiple repositories were not
+  being translated faithfully when creating the manifest file. (#1268)
+
+* Packages installed locally via pak are resolved against configured
+  repositories. (#1305)
+
+* Added support for overriding R package repository resolution behavior. (#1272)
+
+* Push-button publishing from desktop RStudio is now compatible with Connect
+  servers hosted on Snowflake. This includes support for browser-based
+  authentication during deployment. (#1289)
+
+* The `snowflakeConnectionName` parameter now respects the default Snowflake
+  connection name in the `connections.toml` file (when it exists), making it
+  optional in many cases. This is only applicable to Connect servers hosted on
+  Snowflake. (#1283)
+
+* Added support for using identity federation to authenticate against Connect
+  when running in Posit Workbench, when available. This allows deploying to
+  Connect servers without the need to store long-lived credentials. (#1287)
+
+* Upgraded to use `v1` APIs for deploying to Connect servers, which enables
+  new features for specifying settings in the manifest file. (#1280)
+
+* Removed support for log streaming from shinyapps.io due to loss of support
+  for this feature on the shinyapps.io platform (`showLogs(streaming = TRUE)`).
+  If this feature is important to your workflow, please file an issue and we
+  will consider reintroduction of log streaming via rsconnect in Connect Cloud.
+  (#1292)
+
+* Removed several functions, including `addConnectServer()` and
+  `discoverServer()`, as well as HTTP backends other than libcurl,
+  which were deprecated in rsconnect 1.0.0. (#1282)
+
+
 # rsconnect 1.7.0
 
 * Added support for deploying from `manifest.json` files created by
   `writeManifest()`: use the `manifestPath` argument of `deployApp()` and related
   functions to specify the path to an existing manifest file. (#1259)
 
-* `urlEncode()` now uses `curl::curl_escape()` instead of `RCurl::curlEscape()`, 
+* `urlEncode()` now uses `curl::curl_escape()` instead of `RCurl::curlEscape()`,
   as RCurl is a Suggests dependency. (#1265)
-  
+
 * The `User-Agent` header in requests made from rsconnect will now be of the
   format `RSConnect/x.y.z` instead of `rsconnect/x.y.z` in order to satisfy web
   application firewalls that enforce Pascal case.
@@ -54,7 +99,7 @@
 
   Existing Posit Cloud account records may be removed by using
   `removeAccount("yourname", "posit.cloud")`.
-  
+
   Existing Posit Cloud deployment records may be removed by using
   `forgetDeployment(name="deployment", account="yourname", server="posit.cloud")`.